LayoutTests/crypto/workers/subtle/aes-import-jwk-key-expected.txt

[Worker] Test importing a JWK AES key using AES-CBC algorithm in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/aes-import-jwk-key.js
[Worker] Importing a key...
PASS [Worker] crypto.subtle.importKey("jwk", {kty: "RSA"}, "aes-cbc", extractable, ["encrypt", "decrypt", "wrapKey", "unwrapKey"]) rejected promise  with DataError (DOM Exception 30): Data provided to an operation does not meet requirements.
PASS [Worker] key.type is 'secret'
PASS [Worker] key.extractable is true
PASS [Worker] key.algorithm.name is 'AES-CBC'
PASS [Worker] key.algorithm.length is 128
PASS [Worker] key.usages is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
